Converted repo into a Python Package repo.
