+++ cflscrapy/__init__.py
@@ -1,0 +1,7 @@
+##############################################################################
+##
+# Imports
+##
+##############################################################################
+
+from cflscrapy.cfl_scraper import *
